리팩토링: WebGLBuildCopier 서비스 추출
- Editor/Package/WebGLBuildCopier.cs 신설
- CopyWebGLToPublic/CopyAdditionalUserFiles/CopyUserFilesRecursive/GenerateEarlyFetchScript/PrepareAitBuildFolder 5개 이동
- FindSdkRuntimePath 호출은 SdkPathResolver 서비스로 위임
- CopyAdditionalUserFiles/PrepareAitBuildFolder는 internal로 승격 (facade/서비스 간 호출용)
- facade 호출부 전부 서비스 호출로 치환
- 테스트는 E2E 빌드 파이프라인에 위임 (파일 I/O 비중이 커서 유닛 테스트 가치 낮음)
