+++ Editor/AITPackageBuilder.cs
@@ -94,7 +94,7 @@ namespace AppsInToss.Editor
             if (profile == null) profile = AITBuildProfile.CreateProductionProfile();
 
             string buildProjectPath = Path.Combine(projectPath, "ait-build");
-            PrepareAitBuildFolder(buildProjectPath);
+            Package.WebGLBuildCopier.PrepareAitBuildFolder(buildProjectPath);
 
             // npm 경로 확인
             string npmPath = AITNpmRunner.FindNpmPath();
@@ -109,7 +109,7 @@ namespace AppsInToss.Editor
             CopyBuildConfigFromTemplate(buildProjectPath);
 
             Debug.Log("[AIT] Step 2/3: Unity WebGL 빌드 복사 중...");
-            var copyResult = CopyWebGLToPublic(webglPath, buildProjectPath, profile);
+            var copyResult = Package.WebGLBuildCopier.CopyWebGLToPublic(webglPath, buildProjectPath, profile);
             if (copyResult != AITConvertCore.AITExportError.SUCCEED)
                 return (null, copyResult);
 
@@ -165,7 +165,7 @@ namespace AppsInToss.Editor
             if (profile == null) profile = AITBuildProfile.CreateProductionProfile();
 
             string buildProjectPath = Path.Combine(projectPath, "ait-build");
-            PrepareAitBuildFolder(buildProjectPath);
+            Package.WebGLBuildCopier.PrepareAitBuildFolder(buildProjectPath);
 
             // BuildConfig 복사 (WebGL 출력 불필요 — package.json, lockfile 등만)
             Debug.Log("[AIT] [병렬] BuildConfig 파일 복사 중...");
@@ -223,7 +223,7 @@ namespace AppsInToss.Editor
             onProgress?.Invoke(AITConvertCore.BuildPhase.CopyingFiles, 0.15f, "WebGL 빌드 파일 복사 중...");
             Debug.Log("[AIT] [병렬] WebGL 빌드를 ait-build/public으로 복사 중...");
 
-            var copyResult = CopyWebGLToPublic(webglPath, earlyCtx.BuildProjectPath, profile);
+            var copyResult = Package.WebGLBuildCopier.CopyWebGLToPublic(webglPath, earlyCtx.BuildProjectPath, profile);
             if (copyResult != AITConvertCore.AITExportError.SUCCEED)
             {
                 earlyCtx.CancelAndDisposePnpm();
@@ -716,89 +716,6 @@ namespace AppsInToss.Editor
         }
 
         /// <summary>
-        /// 프로젝트 BuildConfig의 추가 파일들을 재귀적으로 복사합니다.
-        /// </summary>
-        internal static void CopyAdditionalUserFiles(string projectBuildConfigPath, string destPath)
-        {
-            if (!Directory.Exists(projectBuildConfigPath)) return;
-
-            // 루트 레벨에서 제외할 파일들
-            var excludeRootFiles = new HashSet<string>
-            {
-                "package.json", "pnpm-lock.yaml", "vite.config.ts",
-                "tsconfig.json", "unity-bridge.ts", "granite.config.ts"
-            };
-
-            // 제외할 폴더들
-            var excludeFolders = new HashSet<string>
-            {
-                "node_modules",
-                ".npm-cache",
-                "dist"
-            };
-
-            CopyUserFilesRecursive(projectBuildConfigPath, destPath, excludeRootFiles, excludeFolders, isRoot: true);
-        }
-
-        /// <summary>
-        /// 재귀적으로 사용자 파일을 복사합니다.
-        /// </summary>
-        private static void CopyUserFilesRecursive(
-            string sourceDir,
-            string destDir,
-            HashSet<string> excludeRootFiles,
-            HashSet<string> excludeFolders,
-            bool isRoot)
-        {
-            // 대상 폴더 생성
-            if (!Directory.Exists(destDir))
-            {
-                Directory.CreateDirectory(destDir);
-            }
-
-            // 파일 복사
-            foreach (var file in Directory.GetFiles(sourceDir))
-            {
-                string fileName = Path.GetFileName(file);
-
-                // 루트 레벨에서만 특정 파일 제외
-                if (isRoot && excludeRootFiles.Contains(fileName))
-                {
-                    continue;
-                }
-
-                string destFile = Path.Combine(destDir, fileName);
-                File.Copy(file, destFile, true);
-
-                // 의미 있는 파일만 로그 출력
-                if (fileName.EndsWith(".ts") || fileName.EndsWith(".tsx") ||
-                    fileName.EndsWith(".js") || fileName.EndsWith(".jsx") ||
-                    fileName.EndsWith(".css") || fileName.EndsWith(".scss"))
-                {
-                    Debug.Log($"[AIT]   ✓ {fileName} (사용자 추가 파일)");
-                }
-            }
-
-            // 하위 폴더 재귀 복사
-            foreach (var dir in Directory.GetDirectories(sourceDir))
-            {
-                string dirName = Path.GetFileName(dir);
-
-                // 제외 폴더 스킵
-                if (excludeFolders.Contains(dirName))
-                {
-                    continue;
-                }
-
-                string destSubDir = Path.Combine(destDir, dirName);
-                CopyUserFilesRecursive(dir, destSubDir, excludeRootFiles, excludeFolders, isRoot: false);
-
-                // 폴더 복사 완료 로그
-                Debug.Log($"[AIT]   ✓ {dirName}/ (사용자 추가 폴더)");
-            }
-        }
-
-        /// <summary>
         /// BuildConfig 템플릿에서 빌드 설정 파일들을 복사합니다.
         /// </summary>
         internal static void CopyBuildConfigFromTemplate(string buildProjectPath)
@@ -876,385 +793,11 @@ namespace AppsInToss.Editor
             }
 
             // 7. 사용자 추가 파일 복사
-            CopyAdditionalUserFiles(projectBuildConfigPath, buildProjectPath);
+            Package.WebGLBuildCopier.CopyAdditionalUserFiles(projectBuildConfigPath, buildProjectPath);
 
             Debug.Log("[AIT] ✓ 빌드 설정 파일 처리 완료");
         }
 
-        /// <summary>
-        /// Unity WebGL 빌드를 public 폴더로 복사합니다.
-        /// </summary>
-        /// <returns>성공 시 SUCCEED, 실패 시 해당 에러 코드</returns>
-        internal static AITConvertCore.AITExportError CopyWebGLToPublic(string webglPath, string buildProjectPath, AITBuildProfile profile = null)
-        {
-            // 프로필이 없으면 기본 프로필 사용
-            if (profile == null)
-            {
-                profile = AITBuildProfile.CreateProductionProfile();
-            }
-
-            var config = UnityUtil.GetEditorConf();
-
-            // Unity WebGL 빌드를 Vite 프로젝트에 복사
-            // - index.html: 프로젝트 루트 (Vite 요구사항)
-            // - Build, TemplateData, Runtime: public 폴더 (정적 자산)
-            string publicPath = Path.Combine(buildProjectPath, "public");
-
-            // public 폴더 생성
-            if (!Directory.Exists(publicPath))
-            {
-                Directory.CreateDirectory(publicPath);
-            }
-
-            // Build 폴더 → public/Build (필수 파일만 선별 복사)
-            string buildSrc = Path.Combine(webglPath, "Build");
-            string buildDest = Path.Combine(publicPath, "Build");
-
-            if (!Directory.Exists(buildSrc))
-            {
-                Debug.LogError(
-                    "[AIT] ✗ 치명적: Build 폴더를 찾을 수 없습니다!\n"
-                    + $"검색 경로: {buildSrc}"
-                );
-                return AITConvertCore.AITExportError.BUILD_FOLDER_MISSING;
-            }
-
-            // Build 폴더에서 실제 파일 이름 찾기
-            // 빌드 마커에서 압축 포맷 정보를 읽어 정확한 확장자로 탐지
-            Debug.Log("[AIT] WebGL 빌드 파일 검색 중...");
-
-            var buildInfo = AITConvertCore.ReadBuildMarker(webglPath);
-            int compressionFormat = buildInfo?.compressionFormat ?? -1;
-            bool decompressionFallback = buildInfo?.decompressionFallback ?? false;
-            var patterns = AITBuildValidator.GetFilePatterns(compressionFormat, decompressionFallback);
-
-            // 폴백 경로 존재 여부: 정확한 패턴(압축 포맷 또는 .unityweb)이 있을 때만 와일드카드 폴백 가능
-            bool hasFallbackPath = compressionFormat >= 0 || decompressionFallback;
-
-            if (buildInfo != null)
-            {
-                string[] formatNames = { "Disabled", "Gzip", "Brotli" };
-                string formatName = compressionFormat >= 0 && compressionFormat < formatNames.Length ? formatNames[compressionFormat] : "Unknown";
-                Debug.Log($"[AIT] 빌드 마커 감지: 압축 포맷 = {formatName} ({compressionFormat}), Decompression Fallback = {decompressionFallback}");
-            }
-
-            // 정확한 패턴으로 시도
-            // 폴백 경로가 있으면 isRequired: false (와일드카드에서 에러 보고)
-            // 폴백 경로가 없으면 isRequired: true (여기서 바로 에러 보고)
-            string loaderFile = AITBuildValidator.FindFileInBuild(buildSrc, patterns["loader"], isRequired: true);
-            string dataFile = AITBuildValidator.FindFileInBuild(buildSrc, patterns["data"], isRequired: !hasFallbackPath);
-            string frameworkFile = AITBuildValidator.FindFileInBuild(buildSrc, patterns["framework"], isRequired: !hasFallbackPath);
-            string wasmFile = AITBuildValidator.FindFileInBuild(buildSrc, patterns["wasm"], isRequired: !hasFallbackPath);
-
-            // 선택적 파일
-            string symbolsFile = AITBuildValidator.FindFileInBuild(buildSrc, patterns["symbols"]);
-
-            // 정확한 패턴으로 못 찾으면 와일드카드로 폴백 (loader는 압축 무관하므로 제외)
-            if (hasFallbackPath)
-            {
-                var fallback = AITBuildValidator.GetFilePatterns(-1);
-                if (string.IsNullOrEmpty(dataFile)) dataFile = AITBuildValidator.FindFileInBuild(buildSrc, fallback["data"], isRequired: true);
-                if (string.IsNullOrEmpty(frameworkFile)) frameworkFile = AITBuildValidator.FindFileInBuild(buildSrc, fallback["framework"], isRequired: true);
-                if (string.IsNullOrEmpty(wasmFile)) wasmFile = AITBuildValidator.FindFileInBuild(buildSrc, fallback["wasm"], isRequired: true);
-                if (string.IsNullOrEmpty(symbolsFile)) symbolsFile = AITBuildValidator.FindFileInBuild(buildSrc, fallback["symbols"]);
-            }
-
-            // 필수 파일 검증
-            var missingFiles = new List<string>();
-            if (string.IsNullOrEmpty(loaderFile)) missingFiles.Add("*.loader.js");
-            if (string.IsNullOrEmpty(dataFile)) missingFiles.Add("*.data");
-            if (string.IsNullOrEmpty(frameworkFile)) missingFiles.Add("*.framework.js");
-            if (string.IsNullOrEmpty(wasmFile)) missingFiles.Add("*.wasm");
-
-            if (missingFiles.Count > 0)
-            {
-                Debug.LogError(
-                    $"[AIT] ✗ 치명적: WebGL 빌드 필수 파일 누락!\n"
-                    + $"누락된 필수 파일: {string.Join(", ", missingFiles)}\n"
-                    + "가능한 원인:\n"
-                    + "  1. Unity WebGL 빌드가 완료되지 않았습니다.\n"
-                    + "  2. WebGL 빌드가 실패했지만 부분 결과물만 남아있습니다.\n"
-                    + "  3. 빌드 설정(압축 방식 등)이 예상과 다릅니다.\n"
-                    + "해결 방법:\n"
-                    + "  1. 'Clean Build' 옵션을 활성화하고 다시 빌드하세요.\n"
-                    + "  2. Unity Console에서 빌드 에러를 확인하세요."
-                );
-                return AITConvertCore.AITExportError.REQUIRED_FILE_MISSING;
-            }
-
-            // Build 대상 폴더 정리 후 재생성
-            if (Directory.Exists(buildDest))
-            {
-                // 실패 시 DeleteDirectory가 내부 경고를 남기지만, 잔존 파일이 이후 복사 단계에
-                // 섞일 수 있으므로 상위 레벨에서도 한 번 더 사용자에게 알림.
-                // 주의: 이 LogWarning은 단순 폴백이 아니라 실제 빌드 오염 위험 신호이므로 Sentry로
-                // 캡처되도록 Warning 레벨을 유지한다. (File.Copy는 덮어쓰지만 복사 대상 목록
-                // (filesToCopy)에 포함되지 않은 잔존 파일은 패키지에 섞여 런타임 오류를 유발할 수 있음)
-                if (!AITFileUtils.DeleteDirectory(buildDest))
-                {
-                    Debug.LogWarning($"[AIT] 이전 빌드 잔여물 정리 실패: {buildDest} — 새 빌드에 오래된 파일이 섞일 수 있습니다");
-                }
-            }
-            Directory.CreateDirectory(buildDest);
-
-            // 필수 파일만 선별 복사
-            var filesToCopy = new List<string> { loaderFile, dataFile, frameworkFile, wasmFile };
-            if (!string.IsNullOrEmpty(symbolsFile))
-            {
-                filesToCopy.Add(symbolsFile);
-            }
-
-            long totalBytes = 0;
-            foreach (var fileName in filesToCopy)
-            {
-                string src = Path.Combine(buildSrc, fileName);
-                string dest = Path.Combine(buildDest, fileName);
-                File.Copy(src, dest, true);
-                UnityUtil.EnsureFileReadable(dest);
-                totalBytes += new FileInfo(src).Length;
-            }
-
-            Debug.Log($"[AIT] ✓ Build 파일 {filesToCopy.Count}개 선별 복사 완료 ({totalBytes / 1024.0 / 1024.0:0.#}MB)");
-
-            // 안전장치: Build/ 폴더에 인식되지 않은 파일이 있으면 로그 출력
-            var allBuildFiles = Directory.GetFiles(buildSrc);
-            var copiedFileNames = new HashSet<string>(filesToCopy);
-            foreach (var file in allBuildFiles)
-            {
-                string name = Path.GetFileName(file);
-                if (!copiedFileNames.Contains(name))
-                {
-                    Debug.Log($"[AIT] Build 폴더에 복사되지 않은 파일: {name}");
-                }
-            }
-
-            // TemplateData 폴더 → public/TemplateData
-            string templateDataSrc = Path.Combine(webglPath, "TemplateData");
-            string templateDataDest = Path.Combine(publicPath, "TemplateData");
-            if (Directory.Exists(templateDataSrc))
-            {
-                UnityUtil.CopyDirectory(templateDataSrc, templateDataDest);
-            }
-
-            // Runtime 폴더 → public/Runtime
-            // 1순위: webgl/ 폴더에 Runtime이 있으면 사용 (AITTemplate 빌드)
-            // 2순위: webgl/ 폴더에 Runtime이 없으면 SDK 템플릿에서 복사
-            string runtimeSrc = Path.Combine(webglPath, "Runtime");
-            string runtimeDest = Path.Combine(publicPath, "Runtime");
-            if (Directory.Exists(runtimeSrc))
-            {
-                UnityUtil.CopyDirectory(runtimeSrc, runtimeDest);
-            }
-            else
-            {
-                // SDK 템플릿에서 Runtime 폴더 복사 (수동 WebGL 빌드 시 AITTemplate 미사용 대응)
-                Debug.LogWarning("[AIT] WebGL 빌드에 Runtime 폴더가 없습니다. SDK 템플릿에서 복사합니다.");
-                Debug.LogWarning("[AIT]    ⚠️ AITTemplate이 아닌 다른 템플릿으로 빌드되었을 수 있습니다.");
-                string sdkRuntimePath = Package.SdkPathResolver.FindSdkRuntimePath();
-                if (!string.IsNullOrEmpty(sdkRuntimePath) && Directory.Exists(sdkRuntimePath))
-                {
-                    UnityUtil.CopyDirectory(sdkRuntimePath, runtimeDest);
-                    Debug.Log("[AIT] ✓ Runtime 폴더: SDK 템플릿에서 복사 완료");
-                }
-                else
-                {
-                    Debug.LogError("[AIT] Runtime 폴더를 찾을 수 없습니다. 'Build And Package'를 사용하세요.");
-                }
-            }
-
-            // StreamingAssets 폴더 → public/StreamingAssets (있는 경우)
-            string streamingAssetsSrc = Path.Combine(webglPath, "StreamingAssets");
-            string streamingAssetsDest = Path.Combine(publicPath, "StreamingAssets");
-            if (Directory.Exists(streamingAssetsSrc))
-            {
-                UnityUtil.CopyDirectory(streamingAssetsSrc, streamingAssetsDest);
-            }
-
-            // index.html → 프로젝트 루트 (Vite가 루트에서 index.html을 찾음)
-            string indexSrc = Path.Combine(webglPath, "index.html");
-            string indexDest = Path.Combine(buildProjectPath, "index.html");
-
-            // index.html 필수 검증
-            if (!File.Exists(indexSrc))
-            {
-                Debug.LogError(
-                    "[AIT] ✗ 치명적: index.html을 찾을 수 없습니다!\n"
-                    + $"검색 경로: {indexSrc}\n"
-                    + "가능한 원인:\n"
-                    + "  1. Unity WebGL 빌드가 완료되지 않았습니다.\n"
-                    + "  2. WebGL 템플릿이 올바르게 설정되지 않았습니다.\n"
-                    + "  3. 이전 빌드가 손상되었습니다.\n"
-                    + "해결 방법:\n"
-                    + "  1. 'Clean Build' 옵션을 활성화하고 다시 빌드하세요.\n"
-                    + "  2. AIT > Clean 메뉴로 빌드 폴더를 삭제 후 재빌드하세요."
-                );
-                return AITConvertCore.AITExportError.INDEX_HTML_MISSING;
-            }
-
-            string indexContent = File.ReadAllText(indexSrc);
-
-            // 프로필 기반 설정값 (Mock 브릿지가 비활성화되면 프로덕션 모드로 간주)
-            string isProduction = profile.enableMockBridge ? "false" : "true";
-            string enableDebugConsole = profile.enableDebugConsole ? "true" : "false";
-
-            // 프로젝트의 index.html에서 사용자 커스텀 섹션 추출 (있는 경우)
-            string projectIndexPath = Path.Combine(Application.dataPath, "WebGLTemplates", "AITTemplate", "index.html");
-            if (File.Exists(projectIndexPath))
-            {
-                string projectIndexContent = File.ReadAllText(projectIndexPath);
-
-                // USER_HEAD 섹션 추출 및 교체
-                string userHeadSection = AITTemplateManager.ExtractHtmlUserSection(projectIndexContent, AITTemplateManager.HTML_USER_HEAD_START, AITTemplateManager.HTML_USER_HEAD_END);
-                if (userHeadSection != null)
-                {
-                    indexContent = AITTemplateManager.ReplaceHtmlUserSection(indexContent, AITTemplateManager.HTML_USER_HEAD_START, AITTemplateManager.HTML_USER_HEAD_END, userHeadSection);
-                    Debug.Log("[AIT] index.html USER_HEAD 섹션 머지됨");
-                }
-
-                // USER_BODY_END 섹션 추출 및 교체
-                string userBodyEndSection = AITTemplateManager.ExtractHtmlUserSection(projectIndexContent, AITTemplateManager.HTML_USER_BODY_END_START, AITTemplateManager.HTML_USER_BODY_END_END);
-                if (userBodyEndSection != null)
-                {
-                    indexContent = AITTemplateManager.ReplaceHtmlUserSection(indexContent, AITTemplateManager.HTML_USER_BODY_END_START, AITTemplateManager.HTML_USER_BODY_END_END, userBodyEndSection);
-                    Debug.Log("[AIT] index.html USER_BODY_END 섹션 머지됨");
-                }
-            }
-
-            // Unity 플레이스홀더 치환
-            indexContent = indexContent
-                .Replace("%UNITY_WEB_NAME%", PlayerSettings.productName)
-                .Replace("%UNITY_WIDTH%", PlayerSettings.defaultWebScreenWidth.ToString())
-                .Replace("%UNITY_HEIGHT%", PlayerSettings.defaultWebScreenHeight.ToString())
-                .Replace("%UNITY_COMPANY_NAME%", PlayerSettings.companyName)
-                .Replace("%UNITY_PRODUCT_NAME%", PlayerSettings.productName)
-                .Replace("%UNITY_PRODUCT_VERSION%", PlayerSettings.bundleVersion)
-                // Unity 표준 URL 형식 (Unity가 치환하지 않은 경우 SDK가 처리)
-                .Replace("%UNITY_WEBGL_LOADER_URL%", $"Build/{loaderFile}")
-                .Replace("%UNITY_WEBGL_DATA_URL%", $"Build/{dataFile}")
-                .Replace("%UNITY_WEBGL_FRAMEWORK_URL%", $"Build/{frameworkFile}")
-                .Replace("%UNITY_WEBGL_CODE_URL%", $"Build/{wasmFile}")
-                .Replace("%UNITY_WEBGL_SYMBOLS_URL%", !string.IsNullOrEmpty(symbolsFile) ? $"Build/{symbolsFile}" : "")
-                // 하위 호환성을 위한 FILENAME 형식 (레거시)
-                .Replace("%UNITY_WEBGL_LOADER_FILENAME%", loaderFile)
-                .Replace("%UNITY_WEBGL_DATA_FILENAME%", dataFile)
-                .Replace("%UNITY_WEBGL_FRAMEWORK_FILENAME%", frameworkFile)
-                .Replace("%UNITY_WEBGL_CODE_FILENAME%", wasmFile)
-                .Replace("%UNITY_WEBGL_SYMBOLS_FILENAME%", symbolsFile)
-                // AIT 커스텀 플레이스홀더
-                .Replace("%AIT_IS_PRODUCTION%", isProduction)
-                .Replace("%AIT_ENABLE_DEBUG_CONSOLE%", enableDebugConsole)
-                .Replace("%AIT_DEVICE_PIXEL_RATIO%", config.devicePixelRatio.ToString())
-                .Replace("%AIT_ICON_URL%", config.iconUrl ?? "")
-                .Replace("%AIT_DISPLAY_NAME%", config.displayName ?? "")
-                .Replace("%AIT_PRIMARY_COLOR%", config.primaryColor ?? "#3182f6")
-                // Early Fetch 스크립트 (로딩 성능 개선)
-                .Replace("%AIT_EARLY_FETCH_SCRIPT%", GenerateEarlyFetchScript(dataFile, wasmFile));
-
-            // 로딩 화면 삽입 (%AIT_LOADING_SCREEN% 플레이스홀더)
-            string loadingContent = "";
-            string projectLoadingPath = AITPackageInitializer.GetProjectLoadingPath();
-
-            // 프로젝트의 loading.html 사용 (SDK 초기화 시 자동 생성됨)
-            if (File.Exists(projectLoadingPath))
-            {
-                loadingContent = File.ReadAllText(projectLoadingPath);
-                Debug.Log("[AIT] ✓ 로딩 화면 적용: " + projectLoadingPath);
-            }
-            else
-            {
-                // 폴백: SDK 기본 템플릿 직접 사용 (초기화가 실행되지 않은 경우)
-                string sdkTemplatePath = AITPackageInitializer.GetSDKLoadingTemplatePath();
-                if (sdkTemplatePath != null)
-                {
-                    loadingContent = File.ReadAllText(sdkTemplatePath);
-                    Debug.Log("[AIT] ✓ SDK 기본 로딩 화면 적용");
-                }
-                else
-                {
-                    Debug.LogWarning("[AIT] 로딩 화면 파일을 찾을 수 없습니다. 빈 로딩 화면이 사용됩니다.");
-                }
-            }
-
-            // %AIT_LOADING_SCREEN% 플레이스홀더 치환
-            indexContent = indexContent.Replace("%AIT_LOADING_SCREEN%", loadingContent);
-
-            File.WriteAllText(indexDest, indexContent, System.Text.Encoding.UTF8);
-            Debug.Log("[AIT] index.html → 프로젝트 루트에 생성");
-
-            // 플레이스홀더 치환 결과 검증
-            if (!AITBuildValidator.ValidatePlaceholderSubstitution(indexContent, indexDest))
-            {
-                return AITConvertCore.AITExportError.PLACEHOLDER_SUBSTITUTION_FAILED;
-            }
-
-            // Runtime/appsintoss-unity-bridge.js 파일도 치환
-            string bridgeSrc = Path.Combine(publicPath, "Runtime", "appsintoss-unity-bridge.js");
-            if (File.Exists(bridgeSrc))
-            {
-                string bridgeContent = File.ReadAllText(bridgeSrc);
-                bridgeContent = bridgeContent.Replace("%AIT_IS_PRODUCTION%", isProduction);
-                File.WriteAllText(bridgeSrc, bridgeContent, System.Text.Encoding.UTF8);
-                Debug.Log($"[AIT] appsintoss-unity-bridge.js Mock 브릿지 모드: {(profile.enableMockBridge ? "활성화" : "비활성화")}");
-            }
-
-            Debug.Log("[AIT] Unity WebGL 빌드 복사 완료");
-            Debug.Log("[AIT]   - index.html → 프로젝트 루트");
-            Debug.Log("[AIT]   - Build, TemplateData, Runtime → public/");
-
-            return AITConvertCore.AITExportError.SUCCEED;
-        }
-
-        /// <summary>
-        /// Unity WebGL 리소스를 조기 fetch하고 Unity loader의 fetch를 인터셉트하는 스크립트를 생성합니다.
-        ///
-        /// 배경: &lt;link rel="preload"&gt;는 CDN의 cache-control: max-age=0 환경에서 동작하지 않습니다.
-        /// preload로 받은 리소스가 즉시 stale 처리되어 Unity loader가 재사용하지 못하고 revalidation 요청을 보내며,
-        /// 이는 이중 로딩과 "preload is found, but is not used" 경고를 유발합니다.
-        ///
-        /// 해결: head에서 JS fetch()를 즉시 시작하고, window.fetch를 일회성으로 인터셉트하여
-        /// Unity loader가 같은 URL을 요청할 때 이미 받은 Response를 반환합니다.
-        /// </summary>
-        private static string GenerateEarlyFetchScript(string dataFile, string wasmFile)
-        {
-            var urls = new System.Collections.Generic.List<string>();
-            if (!string.IsNullOrEmpty(dataFile)) urls.Add($"Build/{dataFile}");
-            if (!string.IsNullOrEmpty(wasmFile)) urls.Add($"Build/{wasmFile}");
-
-            if (urls.Count == 0) return "";
-
-            // JSON 배열로 URL 목록 생성
-            var urlsJson = "[" + string.Join(",", urls.ConvertAll(u => $"\"{u}\"")) + "]";
-
-            return $@"<script>
-    // Early Fetch: HTML 파싱과 동시에 리소스 다운로드를 시작하고,
-    // Unity loader가 같은 URL을 요청할 때 이미 받은 Response를 반환합니다.
-    (function() {{
-        var earlyFetchMap = {{}};
-        var urls = {urlsJson};
-        for (var i = 0; i < urls.length; i++) {{
-            (function(href, fetchUrl) {{
-                var p = fetch(fetchUrl).catch(function() {{ delete earlyFetchMap[href]; return null; }});
-                earlyFetchMap[href] = p;
-            }})(new URL(urls[i], location.href).href, urls[i]);
-        }}
-        var originalFetch = window.fetch;
-        window.fetch = function(resource, init) {{
-            var url = (typeof resource === 'string') ? new URL(resource, location.href).href : resource.url;
-            var pending = earlyFetchMap[url];
-            if (pending) {{
-                delete earlyFetchMap[url];
-                if (Object.keys(earlyFetchMap).length === 0) {{
-                    window.fetch = originalFetch;
-                }}
-                // early fetch 실패 시 null이 반환되므로 원본 fetch로 재시도
-                var self = this, args = arguments;
-                return pending.then(function(r) {{ return r || originalFetch.apply(self, args); }});
-            }}
-            return originalFetch.apply(this, arguments);
-        }};
-    }})();
-    </script>";
-        }
 
         #region Async Packaging
 
@@ -1493,60 +1036,5 @@ namespace AppsInToss.Editor
         }
 
         #endregion
-
-        /// <summary>
-        /// ait-build 폴더 준비 (기존 결과물 정리)
-        /// </summary>
-        private static void PrepareAitBuildFolder(string buildProjectPath)
-        {
-            if (!Directory.Exists(buildProjectPath))
-            {
-                Directory.CreateDirectory(buildProjectPath);
-                Debug.Log("[AIT] ait-build 폴더 생성");
-            }
-            else
-            {
-                Debug.Log("[AIT] 기존 빌드 결과물 정리 중... (node_modules와 설정 파일은 유지)");
-
-                string[] itemsToKeep = new string[]
-                {
-                    "node_modules",
-                    "package.json",
-                    "package-lock.json",
-                    "pnpm-lock.yaml",
-                    "granite.config.ts",
-                    "vite.config.ts",
-                    "tsconfig.json"
-                };
-
-                foreach (string item in Directory.GetFileSystemEntries(buildProjectPath))
-                {
-                    string itemName = Path.GetFileName(item);
-
-                    bool shouldKeep = false;
-                    foreach (string keepItem in itemsToKeep)
-                    {
-                        if (itemName == keepItem)
-                        {
-                            shouldKeep = true;
-                            break;
-                        }
-                    }
-
-                    if (shouldKeep) continue;
-
-                    // SafeDelete/DeleteDirectory는 예외를 던지지 않고 실패 시 내부에서 경고 로그를 남김
-                    if (Directory.Exists(item))
-                    {
-                        AITFileUtils.DeleteDirectory(item);
-                    }
-                    else if (File.Exists(item))
-                    {
-                        AITFileSystemHelper.SafeDelete(item);
-                    }
-                }
-            }
-        }
-
     }
 }
