+++ Editor/Package/WebGLBuildCopier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 03b224c92cbd4e80bf07931a3d656ae8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
